除了插入语句，其他全都搞好了

+++ src/main/java/scc1/java_project/deliver_server/POJO/BillOfLastMonth.java
@@ -8,5 +8,5 @@ import java.math.BigDecimal;
 public class BillOfLastMonth {
     private String payType;
     private Integer count;
-    private BigDecimal number;
+    private Double number;
 }
