+++ src/main/java/scc1/java_project/deliver_server/POJO/ConsumptionOfLastYear.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class ConsumptionOfLastYear {
     private long orderId;
     private String customerName;
-    private long consumption;
+    private double consumption;
 }
